# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Book Reader platform: public reading endpoints\n(books, chapters, verses, search, favorites, bookmarks, notes)\nand role-gated admin endpoints (books CRUD, Excel/CSV import,\nreference data, users, audit log).",
        "title": "Book Reader API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/books/{id}": {
            "get": {
                "tags": [
                    "admin-books"
                ],
                "summary": "Get a book by ID (admin edit form)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BookResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin-books"
                ],
                "summary": "Update a book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BookResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin-books"
                ],
                "summary": "Soft-delete a book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/admin/books/{id}/cover": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "admin-files"
                ],
                "summary": "Upload a book cover image",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "cover image (jpg/png/webp, max 5MB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/admin/books/{id}/pdf": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "admin-files"
                ],
                "summary": "Upload a book's PDF ebook",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "PDF file (max 200MB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BookResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "category": {
                    "$ref": "#/definitions/dto.CategoryResponse"
                },
                "chapters_count": {
                    "type": "integer"
                },
                "cover_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "has_pdf": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "isbn": {
                    "type": "string"
                },
                "language": {
                    "$ref": "#/definitions/dto.LanguageResponse"
                },
                "pdf_url": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "verses_count": {
                    "type": "integer"
                },
                "view_count": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name_en": {
                    "type": "string"
                },
                "name_id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "dto.LanguageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "native_name": {
                    "type": "string"
                }
            }
        },
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                },
                "meta": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "JWT access token, set as an httpOnly cookie by /auth/login.",
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}